+++ src/audio/mixer.rs
@@ -1,6 +1,7 @@
 use crate::audio::commands::InterpolationType;
 use crate::audio::resampler;
 use crate::audio::voice::Voice;
+use crate::debug_log;
 use crate::sequencer::effect::FilterType;
 use crate::sequencer::sample::LoopType;